[FIX] account_reports: safer unlink
closes odoo/enterprise#11236

X-original-commit: 43306757525deaeb0632c59acbe84e0f29b0871f

+++ account_reports/models/account_financial_report.py
@@ -738,7 +738,8 @@ class ReportAccountFinancialReport(models.Model):
         for report in self:
             menu = report.generated_menu_id
             if menu:
-                menu.action.unlink()
+                if menu.action:
+                    menu.action.unlink()
                 menu.unlink()
         return super(ReportAccountFinancialReport, self).unlink()
